refactor: extract EPUB cross-reference evidence
